backend/requirements.txt
django==4.1.7
djangorestframework==3.14.0
djoser==2.1.0
django-filter==22.1
drf-extra-fields==3.4.1
reportlab==3.6.12
psycopg2-binary==2.9.5
django-cors-headers==3.14.0
python-dotenv==1.0.0
gunicorn==20.1.0